Avoid null-references. Add mutex to package manager update.
