+++ src/Orc.NuGetExplorer.Xaml/Services/Interfaces/IPackageCommandService.cs
@@ -18,7 +18,11 @@ namespace Orc.NuGetExplorer
         Task ExecuteAsync(PackageOperationType operationType, IPackageDetails packageDetails, IRepository sourceRepository = null, bool allowedPrerelease = false);
         Task ExecuteAsync(PackageOperationType operationType, IPackageDetails packageDetails);
         Task<bool> CanExecuteAsync(PackageOperationType operationType, IPackageDetails package);
+        
+        [ObsoleteEx(TreatAsErrorFromVersion = "4.5", RemoveInVersion = "5.0")]
         bool IsRefreshRequired(PackageOperationType operationType);
+        [ObsoleteEx(TreatAsErrorFromVersion = "4.5", RemoveInVersion = "5.0")]
+
         string GetPluralActionName(PackageOperationType operationType);
         Task ExecuteInstallAsync(IPackageDetails packageDetails, CancellationToken token);
         Task ExecuteInstallAsync(IPackageDetails packageDetails, IDisposable packageOperationContext, CancellationToken token);
